app update

+++ server/controllers/recordController.js
@@ -46,27 +46,22 @@ router.post("/", async (req, res) => {
   }
 });
 
-router.get("/recs", async (req, res) => {
-  console.log(`---Get My Records-----`)
-  console.log(req.user._id)
-  res.send(`hello`)
-  // try {
-  //   const records = await recordManager.getAllRecordsByOwner(req.user._id);
-  //   console.log(records)
-  //   res.json(records);
-  // } catch (error) {
-  //   return res.json(parser.parseError(error))
-  // }
-});
 
 router.put("/:recordId", async (req, res) => {
   //to DO validations for empty fields
   //const {........} = req.body
-  let isOwner = false;
+  let isOwner = true;
   let currentRecord = await recordManager.getOne(req.params.recordId);
   const recordData = req.body;
 
+  if(currentRecord._ownerId !== req.user._id){
+    isOwner = false
+  }
+
     try {
+    if(!isOwner){
+      throw new Error(`You are not authotized!`)
+    }
     const updatredRecord =  await recordManager.update(req.params.recordId, recordData);
 
       res.json(updatredRecord);
@@ -77,17 +72,20 @@ router.put("/:recordId", async (req, res) => {
 });
 
 router.delete("/:recordId", async (req, res) => {
-  let isOwner = false;
+
+  let isOwner = true;
   let currentRecord = await recordManager.getOne(req.params.recordId);
+  if(currentRecord._ownerId !== req.user._id){
+    isOwner = false
+  }
     try {
+      if(!isOwner){
+        throw new Error(`You are not authotized!`)
+      }
       await recordManager.delete(req.params.recordId);
       res.json({ ok: true });
     } catch (error) {
-        let response = {
-            errors: {},
-            message: error.message,
-          };
-          return response;
+          return res.json(parser.parseError(error))
     }
 });
